test(all): rename nodes
